Refactor wfc_logic to remove visualisation code
